[Refactor] update import statements for ToolUtils to use support package
